refactor: useModal을 범용성있는 훅으로 변경 - 전체 영업 시간 확인용으로만 쓰던 로직을 범용성 있게 변경
